+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 2.1.1
 
-- Adds new alphabet to `Base64Codec`: [bcrypt][https://en.wikipedia.org/wiki/Bcrypt#base64_encoding_alphabet]
+- Adds new alphabet to `Base64Codec`: [bcrypt](https://en.wikipedia.org/wiki/Bcrypt#base64_encoding_alphabet)
 
 # 2.1.0